Renamed `bee.buzz` packages & added autoconfiguration for spring
