+++ bee.buzz/src/main/kotlin/com/beeproduced/bee/buzz/manager/EventError.kt
@@ -1,4 +1,4 @@
-package com.beeproduced.lib.events.manager
+package com.beeproduced.bee.buzz.manager
 
 import com.beeproduced.result.errors.InternalAppError
 import com.beeproduced.result.errors.ResultError
