+++ bee.buzz/src/main/kotlin/com/beeproduced/bee/buzz/manager/Mediator.kt
@@ -1,9 +1,9 @@
-package com.beeproduced.lib.events.manager
+package com.beeproduced.bee.buzz.manager
 
-import com.beeproduced.lib.events.Notification
-import com.beeproduced.lib.events.NotificationHandler
-import com.beeproduced.lib.events.Request
-import com.beeproduced.lib.events.RequestHandler
+import com.beeproduced.bee.buzz.Notification
+import com.beeproduced.bee.buzz.NotificationHandler
+import com.beeproduced.bee.buzz.Request
+import com.beeproduced.bee.buzz.RequestHandler
 import com.beeproduced.result.errors.AppError
 import com.github.michaelbull.result.Result
