+++ bee.buzz/src/simple/kotlin/com/beeproduced/bee/buzz/manager/SimpleEventManager.kt
@@ -1,11 +1,11 @@
-package com.beeproduced.lib.events.manager
+package com.beeproduced.bee.buzz.manager
 
-import com.beeproduced.lib.events.Notification
-import com.beeproduced.lib.events.NotificationHandler
-import com.beeproduced.lib.events.Request
-import com.beeproduced.lib.events.RequestHandler
-import com.beeproduced.lib.events.manager.exceptions.NotificationHandlerNotFound
-import com.beeproduced.lib.events.manager.exceptions.RequestHandlerNotFound
+import com.beeproduced.bee.buzz.Notification
+import com.beeproduced.bee.buzz.NotificationHandler
+import com.beeproduced.bee.buzz.Request
+import com.beeproduced.bee.buzz.RequestHandler
+import com.beeproduced.bee.buzz.manager.exceptions.NotificationHandlerNotFound
+import com.beeproduced.bee.buzz.manager.exceptions.RequestHandlerNotFound
 import com.beeproduced.result.errors.AppError
 import com.github.michaelbull.result.Result
 import org.slf4j.LoggerFactory
