+++ bee.buzz/src/simple/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+com.beeproduced.bee.buzz.autoconfig.BeeBuzzAutoConfiguration
