+++ example/application/src/main/kotlin/com/beeproduced/example/application/DataGenerator.kt
@@ -1,7 +1,7 @@
 package com.beeproduced.example.application
 
 import com.beeproduced.bee.persistent.selection.SimpleSelection
-import com.beeproduced.lib.events.manager.EventManager
+import com.beeproduced.bee.buzz.manager.EventManager
 import com.beeproduced.result.AppResult
 import com.beeproduced.service.media.entities.Film
 import com.beeproduced.service.media.entities.input.CreateFilmInput
