+++ example/application/src/main/kotlin/com/beeproduced/example/application/organisation/service/OrganisationController.kt
@@ -1,7 +1,7 @@
 package com.beeproduced.example.application.organisation.service
 
 import com.beeproduced.bee.persistent.dgs.selection.toDataSelection
-import com.beeproduced.lib.events.manager.EventManager
+import com.beeproduced.bee.buzz.manager.EventManager
 import com.beeproduced.result.extensions.dgs.getDataFetcher
 import com.beeproduced.service.organisation.events.GetAllCompanies
 import com.beeproduced.service.organisation.events.GetAllPersons
