+++ example/application/src/main/kotlin/com/beeproduced/example/application/organisation/service/OrganisationDataLoader.kt
@@ -4,7 +4,7 @@ import com.beeproduced.bee.fetched.annotations.BeeFetched
 import com.beeproduced.bee.fetched.annotations.FetcherIgnore
 import com.beeproduced.bee.persistent.dgs.selection.toDataSelection
 import com.beeproduced.bee.persistent.selection.EmptySelection
-import com.beeproduced.lib.events.manager.EventManager
+import com.beeproduced.bee.buzz.manager.EventManager
 import com.beeproduced.service.organisation.events.GetCompaniesByIds
 import com.beeproduced.service.organisation.events.GetPersonsByIds
 import com.beeproduced.utils.logFor
