+++ example/service.media/events/src/main/kotlin/com/beeproduced/service/media/events/Requests.kt
@@ -2,7 +2,7 @@ package com.beeproduced.service.media.events
 
 import com.beeproduced.bee.persistent.jpa.repository.extensions.PaginationResult
 import com.beeproduced.bee.persistent.selection.DataSelection
-import com.beeproduced.lib.events.Request
+import com.beeproduced.bee.buzz.Request
 import com.beeproduced.service.media.entities.Film
 import com.beeproduced.service.media.entities.input.CreateFilmInput
 import com.beeproduced.service.media.entities.input.FilmPagination
