+++ example/service.media/src/main/kotlin/com/beeproduced/service/media/film.feature/FilmService.kt
@@ -2,7 +2,7 @@ package com.beeproduced.service.media.film.feature
 
 import com.beeproduced.bee.persistent.jpa.repository.extensions.PaginationResult
 import com.beeproduced.bee.persistent.selection.DataSelection
-import com.beeproduced.lib.events.manager.EventManager
+import com.beeproduced.bee.buzz.manager.EventManager
 import com.beeproduced.result.AppResult
 import com.beeproduced.result.errors.BadRequestError
 import com.beeproduced.result.extensions.errors.andThenOnSuccess
