+++ example/service.organisation/src/main/kotlin/com/beeproduced/service/organisation/person/feature/PersonEvents.kt
@@ -1,9 +1,8 @@
 package com.beeproduced.service.organisation.person.feature
 
-import com.beeproduced.lib.events.manager.EventManager
-import com.beeproduced.lib.events.requestHandler
+import com.beeproduced.bee.buzz.manager.EventManager
+import com.beeproduced.bee.buzz.requestHandler
 import com.beeproduced.result.AppResult
-import com.beeproduced.service.organisation.entities.Company
 import com.beeproduced.service.organisation.entities.Person
 import com.beeproduced.service.organisation.events.*
 import com.beeproduced.service.organisation.utils.organisationAdapter
